Upload Dataset
The code automatically fetches this data from Google Drive, so downloading this is not necessary for the project to function. It is included so you can download it if you like. You can also access the data on Google Drive using the links that are commented out in the cell that fetches the data.